+++ DelvUI/Interface/GeneralElements/HUDOptionsConfig.cs
@@ -34,9 +34,9 @@ namespace DelvUI.Interface.GeneralElements
         [Order(20)]
         public bool HideOutsideOfCombat = false;
 
-        [Checkbox("Hide DelvUI outside of duty")]
-        [Order(21)]
-        public bool HideOutsideOfDuty = false;
+        [Checkbox("Show in duty")]
+        [Order(21, collapseWith = nameof(HideOutsideOfCombat))]
+        public bool ShowInDuty = false;
 
         [Checkbox("Hide DelvUI in Gold Saucer")]
         [Order(25)]